fix(xilinx): resolve misra rule 11.3 violations
Fixed below MISRA violation:
- MISRA Violation: MISRA-C:2012 R.11.3:
  - A cast shall not be performed between a pointer to object type and a
    pointer to a different object type.
- Fix:
  - Removed unnecessary cast of pointers.

+++ plat/xilinx/common/include/pm_api_sys.h
@@ -24,7 +24,7 @@
 
 enum pm_ret_status pm_handle_eemi_call(uint32_t flag, uint32_t x0, uint32_t x1,
 				       uint32_t x2, uint32_t x3, uint32_t x4,
-				       uint32_t x5, uint64_t *result);
+				       uint32_t x5, uint32_t *result);
 enum pm_ret_status pm_self_suspend(uint32_t nid,
 				   uint32_t latency,
 				   uint32_t state,
